Upgrade plugin to use Android plugin API v2.
